restructurize the project
